% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquametNWCA-package.r
\docType{package}
\name{aquametNWCA-package}
\alias{aquametNWCA-package}
\title{aquametNWCA: NWCA Metrics and Indices}
\description{
This package calculates vegetation metrics and indices associated with NWCA using field data and trait information. Functions for vegetation are broken into vascular plants, trees, and vegetation types and ground cover. Updated from version 1 by adding increased flexibility in inputs to functions.
}
\keyword{internal}
