% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquametNWCA.r
\docType{package}
\name{aquametNWCA}
\alias{aquametNWCA}
\alias{aquametNWCA-package}
\title{aquametNWCA: A package to calculate metrics used in NWCA 2011}
\description{
aquametNWCA: A package to calculate metrics used in NWCA 2011
}

