% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquametNWCA2.r
\docType{package}
\name{aquametNWCA2}
\alias{aquametNWCA2}
\alias{aquametNWCA2-package}
\title{aquametNWCA2: A package to calculate metrics used in NWCA 2011}
\description{
aquametNWCA2: A package to calculate metrics used in NWCA 2011
}

