% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegType_GrdCover_functions.r
\name{calcBareGround_LitterMets}
\alias{calcBareGround_LitterMets}
\title{Calculate bare ground and litter metrics}
\usage{
calcBareGround_LitterMets(dataIn, nPlot, sampID = "UID")
}
\arguments{
\item{dataIn}{A data frame containing the following variables:
\itemize{
\item sampID: Variable(s) identified in \emph{sampID} argument

\item PLOT (Sample plot from which data were collected)

\item PARAMETER (specific measurement type)

\item RESULT (measured value)
}
The following parameters are used in
calculating tree metrics: 'LITTER_THATCH', 'LITTER_FORB', 
'LITTER_CONIFER', 'LITTER_DECID', 'LITTER_BROADLEAF', 
'LITTER_DEPTH_SW', 'LITTER_DEPTH_NE', 'TOTAL_LITTER', 'WD_FINE', 
'WD_COARSE', 'EXPOSED_SOIL', 'EXPOSED_GRAVEL', 'EXPOSED_ROCK'.
Additional parameters or variables are ignored.}

\item{nPlot}{Number of plots sampled associated with each sample,
including \emph{sampID} variables and NPLOTS.}

\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples}
}
\value{
Either a character string containing an error message when metric
  calculation is not successful, or a data frame. Data frame contains
  \emph{sampID} variables, PARAMETER, RESULT, where values of PARAMETER 
  consist of the metric name concatenated with trait value, with valid values
  of: FREQ_BAREGD, FREQ_EXPOSED_GRAVEL, FREQ_EXPOSED_ROCK, FREQ_EXPOSED_SOIL,
  FREQ_LITTER, FREQ_WD_COARSE, FREQ_WD_FINE, IMP_BAREGD, IMP_EXPOSED_GRAVEL,
  IMP_EXPOSED_ROCK, IMP_EXPOSED_SOIL, IMP_LITTER, IMP_WD_COARSE, IMP_WD_FINE,
  XCOV_BAREGD, XCOV_EXPOSED_GRAVEL, XCOV_EXPOSED_ROCK, XCOV_EXPOSED_SOIL,
  XCOV_LITTER, XCOV_WD_COARSE, XCOV_WD_FINE. A list of metric descriptions is
  provided in the document named VegTypes_GrdCover_Metric_Descriptions.pdf 
  included in the help directory for the package.
}
\description{
Calculate bare ground and litter metrics based on
ground cover data. This function is called by calcVtype_GcovMets().
}
\details{
If any of the parameters are missing, they are assumed to be
zeros (if numeric), and metric values associated with any metrics that
cannot be calculated due to missing parameters are set to a
standardized value.
}
\examples{
head(Vtype_GrCovEx)
# Create data frame with number of plots sampled for each UID
nplots <- data.frame(UID=seq(1:10),NPLOTS=rep(5,10),stringsAsFactors=F)
# alternative approach to creating this data frame
nplots <- plyr::ddply(Vtype_GrCovEx,c('UID'),dplyr::summarise,
NPLOTS=length(unique(PLOT)))

bgEx <- calcBareGround_LitterMets(Vtype_GrCovEx,nplots)

head(bgEx)
unique(bgEx$PARAMETER)
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}
