% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMetricSubsets.r
\name{calcBCmets}
\alias{calcBCmets}
\title{Calculate Bray-Curtis metrics}
\usage{
calcBCmets(indf, sampID = "UID")
}
\arguments{
\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples}

\item{Data}{frame containing cover data summarized by UID, PLOT, and
DISTINCT at the species level. Must also contain at least one of the
following: USDA_NAME (taxon name) or SPECIES_NAME_ID (numeric code
for taxon). If NWCA_NATSTAT is included, a native species version is
also calculated.}
}
\value{
Data frame containing UID, PARAMETER, and RESULT, with one
row of results per parameter and UID. The values for PARAMETER consist
of the metric name concatenated with taxonomic level (represented as
TAXLEVEL below):
\itemize{
\item XBCDIST_SPP: Mean Bray-Curtis distance among plots in sample

\item XBCDIST_NATSPP: Mean Bray-Curtis distance among plots in sample
based only on native species.
}
}
\description{
This function calculates the mean Bray-Curtis
distances among plots for all species, and includes a version using only
native species if the variable NWCA_NATSTAT is included in the input
data frame. This variable is found in the ccNatNWCA dataset.
}
\details{
This function calculates metrics based on all species, and on
native species if the appropriate variable is included in the
input data frame.
}

