% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMetricSubsets.r
\name{calcCategory}
\alias{calcCategory}
\title{Calculate plant category metrics}
\usage{
calcCategory(vascIn, sampID = "UID")
}
\arguments{
\item{vascIn}{Data frame containing cover data summarized by
UID and TAXON, with the following fields:
\itemize{
    \item sampID: Variable(s) identified in \emph{sampID} argument

    \item TAXON: Taxon name

    \item CATEGORY: USDA PLANTS category variable, with valid values
    of DICOT, FERN, GYMNOSPERM, HORSETAIL, LICHEN, LIVERWORT,
    LYCOPOD, MONOCOT, MOSS, or blank

    \item XABCOV: Mean percent cover of taxon across plots

    \item TOTN: Number of taxa in sample

    \item sXRCOV: proportion of summed cover across all taxa
    (XTOTABCOV) represented by taxon in sample

    \item NWCA_NATSTAT (optional): Native status variable with categories
    of 'NAT', 'ADV', 'CRYP', 'INTR', 'UND'
   }}

\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples, 'UID' by default}
}
\value{
Data frame containing \emph{sampID} variables, PARAMETER, RESULT,
  where values of PARAMETER consist of the metric name concatenated with
  trait value (represented as TRAITNM below):
\itemize{
\item N_TRAITNM: Number of taxa with trait

\item PCTN_TRAITNM: Number of taxa with trait as percentage of \emph{TOTN}

\item XABCOV_TRAITNM: Sum of \emph{XABCOV} values across taxa with trait

\item XRCOV_TRAITNM: Sum of \emph{sXRCOV} values across taxa with trait
}
For metrics using native status, the metric name has a suffix
of NAT, ALIEN, AC, INTR, or CRYP.
}
\description{
This function calculates all category-based metrics, including
versions with only native species if the variable NWCA_NATSTAT
is included in the input data frame.
}
\details{
Both CATEGORY and NWCA_NATSTAT variables are recoded to fewer
categories. Taxa with 'UND' as native status are excluded.
}
\examples{
head(VascPlantEx)
exPlant <- prepareData(VascPlantEx)

catEx <- calcCategory(exPlant$byUIDspp)

head(catEx)
unique(catEx$PARAMETER)
}
\references{
US Environmental Protection Agency. 2016. National Wetland
Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US
Environmental Protection Agency, Washington, DC.
}
\author{
Karen Blocksom \email{Blocksom.karen@epa.gov}
}
