% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMetricSubsets.r
\name{calcCC}
\alias{calcCC}
\title{Calculate Wetland Indicator Status metrics}
\usage{
calcCC(indf, sampID = "UID")
}
\arguments{
\item{indf}{Data frame containing cover data summarized by
UID and TAXON, with the following fields:
\itemize{
    \item sampID: Variable(s) identified in sampID argument

    \item TAXON: Taxon name

    \item XABCOV: Mean percent cover of taxon across plots

    \item FREQ: Number plots in which taxon occurs

    \item NWCA_CC: Coefficient of conservatism values by taxon
    from NWCA

    \item Optional: NWCA_NATSTAT: Native status variable with
      categories of 'NAT','ADV','CRYP','INTR','UND'.
      UND taxa are ignored.
   }}

\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples}
}
\value{
Data frame containing sampID variables, PARAMETER, RESULT, where
values of PARAMETER are:
\itemize{
    \item XC: Mean coefficient of conservatism (unweighted)

    \item FQAI: Floral Quality Assessment Index (unweighted)

    \item XC_FREQ: Mean coefficient of conservatism weighted by
    relative frequency

    \item FQAI_FREQ: Floral Quality Assessment Index weighted by
    relative frequency

    \item XC_COV: Mean coefficient of conservatism weighted by
    relative absolute cover

    \item FQAI_COV: Floral Quality Assessment Index weighted by
    relative absolute cover

    \item For metrics based native species, the metric name has a
    suffix of NAT.
   }
}
\description{
This function calculates Wetland Indicator Status (WIS)
metrics, including variations based on native status,
if NWCA_NATSTAT is present in the input data frame.
}
\examples{
head(VascPlantEx)
exPlant <- prepareData(VascPlantEx)

ccEx <- calcCC(exPlant$byUIDspp)

head(ccEx)
unique(ccEx$PARAMETER)
}
\references{
US Environmental Protection Agency. 2016. National Wetland
Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US
Environmental Protection Agency, Washington, DC.
}
\author{
Karen Blocksom \email{Blocksom.karen@epa.gov}
}
