% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMetricSubsets.r
\name{calcDiversity}
\alias{calcDiversity}
\title{Calculate diversity indices}
\usage{
calcDiversity(indf, sampID = "UID")
}
\arguments{
\item{indf}{Data frame containing cover data summarized by
UID and TAXON, with the following fields:
\itemize{
 \item sampID: Variable(s) identified in sampID argument

\item TAXON: Taxon name

\item CATEGORY: USDA PLANTS category variable

\item XABCOV: Mean percent cover of taxon across plots

\item Optional: NWCA_NATSTAT: Native status variable with categories of
'NAT','ADV','CRYP','INTR','UND'
}}

\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples}
}
\value{
Data frame containing sampID variables, PARAMETER, RESULT, where
values of PARAMETER are:
\itemize{
\item D_ALL: Simpson diversity index based on all taxa

\item H_ALL: Shannon-Wiener diversity index based on all taxa

\item J_ALL: Pielou's evenness index based on all taxa
}
If NWCA_NATSTAT is included in input data frame, the
following variations based on subsets of taxa are
calculated:
\itemize{
\item D_NAT: Simpson diversity based only on native species

\item D_ALIEN: Simpson diversity based only on alien species

\item D_AC: Simpson diversity based only on alien and cryptogenic
species

\item H_NAT: Shannon-Wiener diversity index based on native species

\item H_ALIEN: Shannon-Wiener diversity index based on alien species

\item H_AC: Shannon-Wiener diversity index based on alien and
cryptogenic species

\item J_NAT: Pielou's evenness index based on native taxa

\item J_ALIEN: Pielou's evenness index based on alien taxa

\item J_AC: Pielou's evenness index based on alien and cryptogenic taxa
}
}
\description{
Calculate Simpson Diversity, Shannon-Wiener
Diversity, and Pielou's Evenness indices, with variations based on
native status if NWCA_NATSTAT is included in the input data frame.
}
\examples{
head(VascPlantEx)
exPlant <- prepareData(VascPlantEx)

divEx <- calcDiversity(exPlant$byUIDspp)

head(divEx)
unique(divEx$PARAMETER)
}
\references{
US Environmental Protection Agency. 2016. National Wetland
Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US
Environmental Protection Agency, Washington, DC.
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}
