% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMetricSubsets.r
\name{calcNative}
\alias{calcNative}
\title{Calculate metrics based only on native status}
\usage{
calcNative(indf, sampID = "UID")
}
\arguments{
\item{indf}{Data frame containing cover data summarized by
UID and TAXON, with the following fields:
\itemize{
    \item sampID: A character vector containing the name(s) of
variable(s) necessary to identify unique samples

    \item TAXON: Taxon name

    \item XABCOV: Mean percent cover of taxon across plots

    \item TOTN: Number of taxa in sample

    \item sXRCOV: proportion of summed cover across all taxa
    (XTOTABCOV) represented by taxon in sample

    \item sRFREQ: Relative frequency of a taxon, calculated
    as the percentage of the total frequency of taxon
    occurrence across all taxa for a UID

    \item NWCA_NATSTAT: Native status variable with
    categories of 'NAT','ADV','CRYP','INTR','UND'
 }}
}
\value{
Data frame containing sampID variables, PARAMETER, RESULT,
where values of PARAMETER consist of the metric name concatenated
with trait value (represented as TRAITNM below):
\itemize{
  \item PCTN_TRAITNM: Number of taxa with trait as percentage of TOTN

  \item XABCOV_TRAITNM: Sum of XABCOV values across taxa with trait

  \item XRCOV_TRAITNM: Sum of sXRCOV values across taxa with trait

  \item RFREQ_TRAITNM: Sum of sRFREQ values across taxa with trait value

  \item RIMP_TRAITNM: Relative importance ((RFREQ_TRAITVAL + XRCOV_TRAITVAL)/2)
  of taxa with trait value
  }
}
\description{
This function calculates all metrics based
only on native status.
}
\examples{
head(VascPlantEx)
exPlant <- prepareData(VascPlantEx)

natEx <- calcCC(exPlant$byUIDspp)

head(natEx)
unique(natEx$PARAMETER)
}
\author{
Karen Blocksom \email{Blocksom.karen@epa.gov}
}
\references{
US Environmental Protection Agency. 2016. National Wetland
Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US
Environmental Protection Agency, Washington, DC.
}

