% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMetricSubsets.r
\name{calcRichness}
\alias{calcRichness}
\title{Calculate richness metrics}
\usage{
calcRichness(byUIDspp, byPlotspp, byUIDgen, byPlotgen, byUIDfam, byPlotfam)
}
\arguments{
\item{byUIDspp}{Data frame containing species data summarized to the UID
and TAXON level, with the following variables:
\itemize{
\item UID: Unique sample identifier

\item STATE: State two-letter abbreviation for site

\item USAC_REGION: U.S. Army Corps of Engineers region name for site

\item TAXON: Taxon name

\item COVER: Total percent cover for taxon

\item DISTINCT: Distinctness (0/1) of taxon within sample, where
taxa above species level are not distinct (0) if taxa
at a lower level are also included in sample. 
}}

\item{byPlotspp}{Data frame containing species data summarized
to the UID, PLOT, and TAXON level, with the following variables:
\itemize{
\item UID: Unique sample identifier

\item PLOT: Plot from which data were collected

\item STATE: State two-letter abbreviation for site

\item USAC_REGION: U.S. Army Corps of Engineers region name for site

\item TAXON: Taxon name

\item COVER: Total percent cover for taxon

\item DISTINCT: Distinctness (0/1) of taxon within sample, where
taxa above species level are not distinct (0) if taxa
at a lower level are also included in sample.
}}

\item{byUIDgen}{Data frame containing genus-level data summarized
to the UID and TAXON level and containing the same variables as
byUIDspp (described above).}

\item{byPlotgen}{Data frame containing genus-level data
summarized to the UID, PLOT, and TAXON level and containing the
same variables as byPlotspp (described above).}

\item{byUIDfam}{Data frame containing family-level data
summarized to the UID and TAXON level and containing the same
variables as byUIDspp (described above).}

\item{byPlotfam}{Data frame containing family-level data
summarized to the UID, PLOT, and TAXON level and containing the
same variables as byPlotspp (described above).}
}
\value{
Data frame containing UID, PARAMETER, and RESULT, with one row of
results per parameter and UID. The values for PARAMETER consist of the
metric name concatenated with taxonomic level (represented as SPP, GEN,
and FAM below):
\itemize{
\item TOTN_SPP, TOTN_GEN, TOTN_FAM: Number of unique taxa in sample (UID)

\item XN_SPP, XN_GEN, XN_FAM: Mean number of taxa per plot

\item MEDN_SPP, MEDN_GEN, MEDN_FAM: Median number of taxa per plot

\item SDN_SPP, SDN_GEN, SDN_FAM: Standard deviation of number of taxa per plot

\item N_PLOTS: Number of plots sampled for UID
}
If NWCA_NATSTAT is present in the input data frame, the following
metrics are calculated. GRP in the metric names below represents
native status values of NAT, ADV, CRYP, INTR, and subsets ALIEN
(INTR + ADV) and AC (ALIEN and CRYP):
\itemize{
\item TOTN_GRP: Number of unique taxa in sample (UID)

\item XN_GRP: Mean number of taxa per plot

\item MEDN_GRP: Median number of taxa per plot

\item SDN_GRP: Standard deviation of number of taxa per plot
}
}
\description{
This function calculates richness metrics using plot- and
UID-based datasets at the species, genus, and family levels.
}
\details{
The prepareData() function creates a list object with
all of the necessary input data frames. For each taxonomic level,
the function createDFs() creates a list with a data frame summarized
by UID and one by UID and PLOT.
}
\examples{
  head(VascPlantEx)
  exPlant <- prepareData(VascPlantEx)

  richEx <- calcRichness(exPlant$byUIDspp,exPlant$byPlotspp,
         exPlant$byUIDgen,exPlant$byPlotgen,exPlant$byUIDfam,exPlant$byPlotfam)

  head(richEx)
  unique(richEx$PARAMETER)
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}
\references{
US Environmental Protection Agency. 2016. National
Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006.
US Environmental Protection Agency, Washington, DC.
}

