% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMetric_functions.r
\name{calcTreeCntMets}
\alias{calcTreeCntMets}
\title{Calculate tree count metrics}
\usage{
calcTreeCntMets(treeIn, nPlot, sampID = "UID")
}
\arguments{
\item{treeIn}{A data frame containing the following variables:
\itemize{
\item sampID: Variable(s) identified in \emph{sampID} argument

\item PLOT: Sample plot from which data were collected

\item PARAMETER: specific measurement type

\item RESULT: measured value
}
The following parameters are used in calculating tree metrics:
'XXTHIN_TREE', 'XTHIN_TREE', 'THIN_TREE', 'JR_TREE', 'THICK_TREE', 
'XTHICK_TREE', 'XXTHICK_TREE'. Additional parameters or variables
are ignored.}

\item{nPlot}{A data frame with the 
number of plots sampled associated with
each sample, with \emph{sampID} variables and NPLOTS.}

\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples, 'UID' by default}
}
\value{
Either a character string containing an error message when metric
  calculation is not successful, or a data frame. The data frame contains the
  \emph{sampID} variables, PARAMETER, RESULT, where PARAMETER values include:
  TOTN_TREES, XN_TREES, TOTN_JR_TREE, TOTN_THICK_TREE, TOTN_THIN_TREE, 
  TOTN_XTHICK_TREE, TOTN_XTHIN_TREE, TOTN_XXTHICK_TREE, TOTN_XXTHIN_TREE, 
  XN_JR_TREE, XN_THICK_TREE, XN_THIN_TREE, XN_XTHICK_TREE, XN_XTHIN_TREE, 
  XN_XXTHICK_TREE, XN_XXTHIN_TREE. A list of metric descriptions is provided
  in the document named Tree_Metric_Descriptions.pdf included in the help
  directory for the package.
}
\description{
This function calculates tree count metrics
using tree data, and is called by function calcTreeMetrics()
}
\details{
If any of the parameters are missing, they are assumed to be
zeros, and metric values associated with any metrics that cannot be
calculated due to missing parameters are set to 0.
}
\examples{
head(TreesEx)
nplots <- plyr::ddply(TreesEx,c('UID'),dplyr::summarise
,NPLOTS=length(unique(PLOT)))

tcntEx <- calcTreeCntMets(TreesEx, nPlot=nplots, sampID='UID')

head(tcntEx)
unique(tcntEx$PARAMETER)
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}
