% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTreeCoverMets.r
\name{calcTreeCoverMets}
\alias{calcTreeCoverMets}
\title{Calculate tree cover metrics}
\usage{
calcTreeCoverMets(treeIn, nPlot, sampID = "UID")
}
\arguments{
\item{treeIn}{A data frame containing the following variables:
\itemize{
\item sampID: Variable(s) identified in \emph{sampID} argument

\item PAGE: Page from field form

\item LINE: Line number from field form

\item PLOT: Sample plot from which data were collected

\item PARAMETER: specific measurement type

\item RESULT: measured value
}
The following parameters are used in
calculating tree metrics: 'VSMALL_TREE', 'SMALL_TREE', 'LMED_TREE'
, 'HMED_TREE', 'TALL_TREE', 'VTALL_TREE'.  Additional parameters
or variables are ignored.}

\item{nPlot}{A data frame with the
number of plots sampled associated with
each sample, with \emph{sampID} variables and NPLOTS.}

\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples, 'UID' by default}
}
\value{
Either a character string containing an error message when metric
  calculation is not successful, or a data frame. The data frame contains the
  \emph{sampID} variables, PARAMETER, RESULT, where PARAMETER values include:
  N_TREESPP, N_TALL_TREE, N_HMED_TREE, N_LMED_TREE, N_SMALL_TREE,
  N_VSMALL_TREE, N_VTALL_TREE, N_TREE_UPPER, N_TREE_MID, N_TREE_GROUND,
  PCTN_TREE_UPPER, PCTN_TREE_MID, PCTN_TREE_GROUND. A list of metric
  descriptions is provided in the document named \href{https://github.com/USEPA/aquametNWCA/blob/main/inst/Tree_Metric_Descriptions.pdf}{Tree_Metric_Descriptions.pdf}
  included in the help directory for the package.
}
\description{
This function calculates tree cover metrics
using tree data, and is called by function calcTreeMetrics().
}
\details{
If any of the parameters are missing, they are assumed
to be zeros, and metric values associated with any metrics that
cannot be calculated due to missing parameters are set to 0.
}
\examples{
head(TreesEx)
nplots <- aggregate(
  x = list(NPLOTS = TreesEx$PLOT), by = TreesEx[c("UID")],
  FUN = function(x) {
    length(unique(x))
  }
)

tcvrEx <- calcTreeCoverMets(TreesEx, nPlot = nplots, sampID = "UID")

head(tcvrEx)
unique(tcvrEx$PARAMETER)
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}
