% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTreeMetrics.r
\name{calcTreeMets}
\alias{calcTreeMets}
\title{Calculate NWCA 2011 tree metrics based on data collected from
field form V-4a}
\usage{
calcTreeMets(treeIn, sampID = "UID")
}
\arguments{
\item{treeIn}{A data frame containing the following variables:
\itemize{
   \item sampID - A character vector containing the name(s) of
variable(s) necessary to identify unique samples

   \item PAGE: Page number from field form

   \item LINE: Line number from field form

   \item PLOT: Sample plot from which data were collected

   \item PARAMETER: specific measurement type

   \item RESULT: measured value
}
The following parameters are used in
calculating tree metrics: 'XXTHIN_SNAG','XTHIN_SNAG','THIN_SNAG',
'JR_SNAG','THICK_SNAG','XTHICK_SNAG','XXTHICK_SNAG','XXTHIN_TREE',
'XTHIN_TREE','THIN_TREE','JR_TREE','THICK_TREE','XTHICK_TREE',
'XXTHICK_TREE','VSMALL_TREE','SMALL_TREE','LMED_TREE','HMED_TREE',
'TALL_TREE','VTALL_TREE'. Additional parameters or variables are
ignored.}
}
\value{
Either a character string containing an error message when 
metric calculation is not successful, or a data frame. The first columns 
of the data frame are the sampID variables and subsequent columns are 
named for each metric and contain metric values. A list of metrics
is provided in the document named "Tree_Metric_Descriptions.pdf" included 
in the help directory for the package.
}
\description{
This function calculates the NWCA 2011 tree metrics.
It assumes input data are organized in long format so that each
row represents the value for a single parameter value within a single
plot at a site.
}
\details{
If any of the parameters are missing, they are assumed to be
zeros, and metric values associated with any metrics that cannot be
calculated due to missing parameters are set to 0.
}
\examples{
head(TreesEx)

exOut <- calcTreeMets(TreesEx)

head(exOut)
str(exOut)
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}
\references{
US Environmental Protection Agency. 2016. National
Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006.
US Environmental Protection Agency, Washington, DC.
}

