% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVMMI_2016_fromMets.r
\name{calcVMMI_2016_fromMets}
\alias{calcVMMI_2016_fromMets}
\title{Calculate Vegetation MMI and assign condition class}
\usage{
calcVMMI_2016_fromMets(metsIn, sampID = "UID", wetcls_grp = NULL)
}
\arguments{
\item{metsIn}{Data frame containing, at a minimum: 
\itemize{ 
  \item sampID: Variables identified in \emph{sampID} argument
  
  \item wetcls_grp: Variable identified in \emph{wetcls_grp} argument, 
  valid values of this variable are: EH, EW, PRLH, PRLW
  
  \item Metrics required, depending on wetland class 
  groups in data: 
   \itemize{
    \item EH: N_ANNUAL, PCTN_NATSPP, XRCOV_FORB, XRCOV_HTOL,
    XRCOV_MONOCOTS_NAT, XRCOV_SEN
 
    \item EW: PCTN_ISEN, PCTN_MONOCOT, PCTN_NATSPP, RIMP_NATSPP,
    XCOV_WD_FINE, XRCOV_GRAMINOID
    
    \item PRLH: FQAI_ALL, N_TOL, PCTN_OBL_FACW, XRCOV_NATSPP
    
    \item RFREQ_NATSPP, XC_ALL, XRCOV_MONOCOTS_NAT, XRCOV_NATSPP
    
   }
  \item RPT_UNIT (optional): NWCA 2016 reporting units with the following 
  valid values: ARW, ATL, GFC, GPL, ICP, NCE, PAC, SAP, TPL, WVM. 
  Required to assign condition class based on VMMI_2016.
  }}

\item{sampID}{A character vector containing the name(s) of variable(s)
necessary to identify unique samples, 'UID' by default}

\item{wetcls_grp}{String containing name of variable containing wetland 
class group, with valid values of EH (estuarine herbaceous),
EW (estuarine woody), PRLH (inland herbaceous), and 
PRLW (inland woody).}
}
\value{
Data frame containing: \itemize{ 
\item \emph{sampID} Variable(s)
  found in the argument sampID
  
  \item All metrics used in wetland class-specific MMIs, with 
  _SC16 on the end of the metric
  
  \item VMMI_2016: NWCA 2016 Vegetation Multimetric Index on 
  100-point scale } If
  RPT_UNIT is provided, the following is also provided as output: 
  \itemize{ 
  \item VEGCOND_2016: Vegetation condition class, based on VMMI score
  and site location, as assigned for NWCA 2016. Valid values are Good, Fair,
  and Poor. }
}
\description{
This function calculates the NWCA 2016 Vegetation
Multimetric Index (VMMI) from metric and wetland class group 
inputs. If the appropriate
variable describing NWCA 2016 reporting units (RPT_UNIT)
is included in the input data frame,
condition class (Good/Fair/Poor) will also be assigned.
}
\examples{
head(vmmiMetEx)

# Test using ECO_X_WETGRP variable in data frame
check.1 <- calcVMMI_fromMets(vmmiMetEx)

# Now drop ECO_X_WETGRP and use NWCA_ECO4 and NWCA_WET_GRP
check.2 <- calcVMMI_fromMets(subset(vmmiMetEx,select=-ECO_X_WETGRP))

# Now run without ECO_X_WETGRP, NWCA_ECO4, and NWCA_WET_GRP
check.3 <- calcVMMI_fromMets(subset(vmmiMetEx,select=c('UID','FQAI_ALL'
                          ,'N_TOL','RIMP_NATSPP','XRCOV_MONOCOTS_NAT')))
}
\references{
US Environmental Protection Agency. 2016. National Wetland
  Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US
  Environmental Protection Agency, Washington, DC.
}
\author{
Karen Blocksom  \email{blocksom.karen@epa.gov}
}
