% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMetricSubsets.r
\name{calcVMMImets}
\alias{calcVMMImets}
\title{Calculate only metrics used in NWCA 2011 VMMI}
\usage{
calcVMMImets(indf, sampID = "UID")
}
\arguments{
\item{indf}{Data frame containing cover data summarized at the UID
and TAXON level:
\itemize{
 \item sampID: Variable(s) identified in sampID argument

\item TAXON: Taxon name

\item NWcA_CC: Coefficient of Conservatism as used in NWCA, valid
values range between 0 and 10

\item NWCA_NATSTAT: Native status as used in NWCA, with native taxa
indicated by 'NAT'

\item CATEGORY: Category abbreviations from USDA PLANTS database,
with monocots indicated by 'MONOCOT'

\item sXRCOV: Proportion of summed cover across all taxa
(XTOTABCOV) represented by taxon in sample

\item sRFREQ: Relative frequency of a taxon, calculated as the
percentage of the total frequency of taxon occurrence across
all taxa for a UID.
}}

\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples}
}
\value{
Data frame containing sampID variables, PARAMETER, RESULT, with the
following PARAMETER values:
\itemize{
\item FQAI_ALL: Floristic Quality Assessment Index, based on all
species

\item N_TOL: Number of tolerant species, as defined by C-value

\item RIMP_NATSPP: Relative importance of native species

\item XRCOV_MONOCOTS_NAT: Mean relative cover of native monocots
}
RESULT contains the metric value for each parameter.
}
\description{
This function calculates FQAI_ALL, N_TOL, RIMP_NATSPP,
and XRCOV_MONOCOTS_NAT metrics, which are used in the NWCA 2011
Vegetation Multimetric Index (VMMI).
}
\details{
To calculate the VMMI as used in NWCA 2011, the default
taxa lists must be used to create the input data frame.
}
\examples{
head(VascPlantEx)
exPlant <- prepareData(VascPlantEx)

vmmiEx <- calcVMMImets(exPlant$byUIDspp)

head(vmmiEx)
unique(vmmiEx$PARAMETER)
}
\references{
US Environmental Protection Agency. 2016.
National Wetland Condition Assessment: 2011 Technical Report.
EPA-843-R-15-006. US Environmental Protection Agency,
Washington, DC.
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}
