% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVegType_GrdCvr_Metrics.r
\name{calcVtype_GcovMets}
\alias{calcVtype_GcovMets}
\title{Calculate NWCA 2011 vegetation types and ground surface
attribute metrics based on data collected from field form V-3}
\usage{
calcVtype_GcovMets(dataIn, nPlotIn, sampID = "UID", survyear = "2011")
}
\arguments{
\item{dataIn}{A data frame containing the following variables:
\itemize{
\item sampID: Variables identified in the \emph{sampID} argument

\item PLOT: Sample plot from which data were collected

\item PARAMETER: specific measurement type

\item RESULT: measured value
}
The following parameters are used in
calculating tree metrics: 'SANDT_CLASS','SUBMERGED_AQ',
'FLOATING_AQ','LIANAS','VTALL_VEG','TALL_VEG','HMED_VEG',
'MED_VEG','SMALL_VEG','VSMALL_VEG','PEAT_MOSS','BRYOPHYTES',
'LICHENS','ARBOREAL','ALGAE','MACROALGAE','TIME','MINIMUM_DEPTH',
'MAXIMUM_DEPTH','PREDOMINANT_DEPTH','TOTAL_WATER','WATER_NOVEG',
'WATER_AQVEG','WATER_EMERGVEG','LITTER_THATCH','LITTER_FORB',
'LITTER_CONIFER','LITTER_DECID','LITTER_BROADLEAF',
'LITTER_DEPTH_SW','LITTER_DEPTH_NE','TOTAL_LITTER','WD_FINE',
'WD_COARSE','EXPOSED_SOIL','EXPOSED_GRAVEL','EXPOSED_ROCK',
'EXPOSED_SOIL','EXPOSED_GRAVEL','EXPOSED_ROCK','WD_FINE',
'WD_COARSE','TOTAL_LITTER'. Additional parameters or variables
are ignored.}

\item{nPlotIn}{A data frame with the
number of plots sampled associated with
each sample, with \emph{sampID} variables and NPLOTS.}

\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples, 'UID' by default}

\item{survyear}{A string with the survey year. Some parameters only
measured in 2011, applies to water cover metrics and bare ground
and litter metrics.}
}
\value{
Either a character string containing an error message when metric
  calculation is not successful, or a data frame. The first column of the
  data frame is the \emph{sampID} variables and subsequent columns are named
  for each metric and contain metric values. A list of vegetation type and
  ground cover metrics is installed with the package. Use
  system.file("VegTypes_GrdCover_Metric_Descriptions.pdf", package="aquametNWCA")
  to locate the file.
}
\description{
This function calculates the NWCA 2011 vegetation types and
ground surface attribute metrics.  It assumes input data are organized
so that each row represents the cover for a single taxon within a single
plot at a site. It also assumes a single column is used to identify
each unique sample. Missing values and blanks should be removed from the
input data frame.
}
\details{
If any of the parameters are missing, they are assumed to be
zeros (if numeric), and metric values associated with any metrics that
cannot be calculated due to missing parameters are set to a standardized
value.
}
\examples{
head(Vtype_GrCovEx)

subVtype_GrCovEx <- subset(Vtype_GrCovEx, PARAMETER != "SANDT_CLASS")

nplots <- aggregate(
  x = list(NPLOTS = subVtype_GrCovEx$PLOT),
  by = subVtype_GrCovEx[c("UID")],
  FUN = function(x) {
    length(unique(x))
  }
)

exOut <- calcVtype_GcovMets(Vtype_GrCovEx, nPlotIn = nplots, sampID = "UID")

str(exOut)
head(exOut)
}
\references{
US Environmental Protection Agency. 2016. National Wetland
  Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US
  Environmental Protection Agency, Washington, DC.
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}
