% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWcovMets.r
\name{calcWcovMets}
\alias{calcWcovMets}
\title{Calculate water cover metrics}
\usage{
calcWcovMets(dataIn, nPlot, sampID = "UID", survyear = "2011")
}
\arguments{
\item{dataIn}{A data frame containing the following variables:
\itemize{
\item sampID: Variable(s) identified in \emph{sampID} argument

\item PLOT: Sample plot from which data
were collected

\item PARAMETER: specific measurement type

\item RESULT: measured value
}
The following parameters are used in
calculating water cover metrics: 'TIME', 'MINIMUM_DEPTH',
'MAXIMUM_DEPTH', 'PREDOMINANT_DEPTH', 'TOTAL_WATER',
'WATER_NOVEG', 'WATER_AQVEG', 'WATER_EMERGVEG'.
Additional parameters or variables are ignored.
WATER_NOVEG, WATER_AQVEG, MINIMUM_DEPTH, and
MAXIMUM_DEPTH not measured in 2016.}

\item{nPlot}{A data frame with the
number of plots sampled associated with each
sample with \emph{sampID} variables and NPLOTS}

\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples,
'UID' by default.}

\item{survyear}{A string with the survey year. The default is '2011'.
Starting in 2016, only TIME, TOTAL_WATER, and PREDOMINANT_DEPTH is
measured.}
}
\value{
Either a character string containing an error message when metric
  calculation is not successful, or a data frame. Data frame contains
  \emph{sampID} variables, PARAMETER, RESULT, where values of PARAMETER
  consist of the metric name concatenated with trait value, with valid values
  of: MIN_H2O_DEPTH, MAX_H2O_DEPTH, XH2O_DEPTH_AA, MIN_COV_H2O, MAX_COV_H2O,
  FREQ_H2O, FREQ_H2O_AQVEG, FREQ_H2O_EMERGVEG, FREQ_H2O_NOVEG, XCOV_H2O,
  XCOV_H2O_AQVEG, XCOV_H2O_EMERGVEG, XCOV_H2O_NOVEG, IMP_H2O, IMP_H2O_AQVEG,
  IMP_H2O_EMERGVEG, IMP_H2O_NOVEG, XH2O_DEPTH. A list of metric descriptions
  is provided in the document named \href{https://github.com/USEPA/aquametNWCA/blob/main/inst/VType_GrdCvr_Metric_Descriptions.pdf}{VegTypes_GrdCover_Metric_Descriptions.pdf}
  included in the help directory for the package.
}
\description{
This function calculates metrics water
cover and depth based on ground cover data. This function is
called by calcVtype_GcovMets().
}
\details{
If any of the parameters are missing, they are assumed
to be zeros (if numeric), and metric values associated with any
metrics that cannot be calculated due to missing parameters are set
to a standardized value.
}
\examples{
head(Vtype_GrCovEx)
# Create data frame with number of plots sampled for each sampID
nplots <- data.frame(UID = seq(1:10), NPLOTS = rep(5, 10), stringsAsFactors = F)
# alternative approach to creating this data frame
nplots <- aggregate(
  x = list(NPLOTS = Vtype_GrCovEx$PLOT),
  by = Vtype_GrCovEx[c("UID")],
  FUN = function(x) {
    length(unique(x))
  }
)

wcovEx <- calcWcovMets(Vtype_GrCovEx, nplots)

head(wcovEx)
unique(wcovEx$METRIC)
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}
