% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VascPlantMetric_functions.r
\name{createDFs}
\alias{createDFs}
\title{Create data frames for function input}
\usage{
createDFs(tvar, indf, taxa)
}
\arguments{
\item{tvar}{String with the level of taxonomy
('USDA_NAME','GENUS','FAMILY')}

\item{indf}{Data frame with vegetation cover data, having
the following variables:
\itemize{

\item UID - Unique sample identifier

\item PLOT - Plot number of data (1 to 5 possible)

\item STATE - State postal code of site

\item USAC_REGION - U.S. Army Corps of Engineers Region Name

\item USDA_NAME - Taxon name, must match with taxa data frame

\item COVER - Percentage estimated vegetation cover by TAXON in PLOT
}}

\item{taxa}{Data frame containing USDA_NAME, GENUS, CATEGORY,
GROWTH_HABIT, and DURATION variables. Dataset taxaNWCA is the
default if not specified. These variables are assumed to be
populated with abbreviations as found in UsDA PLANTS database.}
}
\value{
A list containing two data frames. The first data frame
summarizes data by UID and TAXON and contains UID, STATE,
USAC_REGION, TAXON, NUM, XABCOV, DISTINCT, and NPLOTS. NUM is the
number of plots in which taxon occurs, and XABCOV is the mean
absolute COVER across plots. DISTINCT is the value 1 assigned to
each row. NPLOTS is the number of plots sampled.

The second summarizes by UID, PLOT, and TAXON. Each data frame
contains UID, PLOT, STATE, USAC_REGION, TAXON, COVER, and
DISTINCT. DISTINCT assigns the value for each taxon as 1 if the
taxon has COVER>0 or 0 if not. COVER is the sum of the COVER
variable by UID, PLOT, and TAXON.
}
\description{
This internal function merges the inputs of
plant cover data and taxalist and summarizes by the taxonomic
level specified. Not intended for use on its own.
}
\examples{
head(VascPlantEx)
data(taxaNWCA)

outEx <- createDFs('GENUS',VascPlantEx,taxaNWCA)
head(outEx$byUID)
head(outEx$byPlot)
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}

