% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VascPlantMetric_functions.r
\name{int.calcTraits_Indicator}
\alias{int.calcTraits_Indicator}
\title{Calculate metrics using traits with only two values (0/1)}
\usage{
int.calcTraits_Indicator(vascIn, trait, sampID)
}
\arguments{
\item{vascIn}{Input data frame with \emph{sampID} variables, TAXON, TOTN,
XABCOV, sXRCOV, and a binary variable with name in \emph{trait} argument 
and possible values of 0 and 1, with 1 indicating trait present for taxon.}

\item{trait}{Character string with name of variable containing traits of
interest}

\item{sampID}{A character vector containing the name(s) of variable(s)
necessary to identify unique samples}
}
\value{
Data frame containing \emph{sampID} variables, PARAMETER, RESULT,
  where values of PARAMETER consist of the metric name concatenated with
  trait name (represented as TRAITNM below):

N_TRAITNM: Number of taxa with trait

PCTN_TRAITNM: Number of taxa with trait as percentage of \emph{TOTN}

XABCOV_TRAITNM: Sum of \emph{XABCOV} values across taxa with trait

XRCOV_TRAITNM: Sum of \emph{sXRCOV} values across taxa with trait
}
\description{
This internal function calculates metrics for traits
that are indicator values (0 or 1). Output feeds into 
\code{int.combTraits()} function. Used in 
\code{calcWIS()} function.
}
\section{Warning}{
 This function not intended for use on its own
}

\author{
Karen Blocksom
}
