% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VascPlantMetric_functions.r
\name{int.calcTraits_MultCat.alt}
\alias{int.calcTraits_MultCat.alt}
\title{Alternate metric calculations for traits with >2 categories}
\usage{
int.calcTraits_MultCat.alt(vascIn, trait, sampID)
}
\arguments{
\item{vascIn}{Input data frame with \emph{sampID} variables, TAXON, TOTN,
XABCOV, sXRCOV, sRFREQ, and variable with name in argument trait.
\emph{TOTN} is the total number of taxa at the lowest level for
\emph{sampID} variables. \emph{XABCOV} is the mean absolute cover of taxon.
\emph{sXRCOV} is the percentage of the total sum of absolute cover across
all taxa for a sample. \emph{sRFREQ} is the relative frequency of a taxon,
calculated as the percentage of the total frequency of taxon occurrence
across all taxa for a sample.}

\item{trait}{Character string with name of variable containing traits of
interest}

\item{sampID}{A character vector containing the name(s) of variable(s)
necessary to identify unique samples}
}
\value{
Data frame containing \emph{sampID} variables, PARAMETER, RESULT,
  where values of PARAMETER consist of the metric name concatenated with each
  value of the trait (represented as TRAITVAL below):

  N_TRAITVAL: Number of taxa with trait value

  PCTN_TRAITVAL: Number of taxa with trait value as percentage of TOTN

  XABCOV_TRAITVAL: Sum of XABCOV values across taxa with trait value

  XRCOV_TRAITVAL: Sum of sXRCOV values across taxa with trait value

  RFREQ_TRAITVAL: Sum of sRFREQ values across taxa with trait value

  RIMP_TRAITVAL: Relative importance ((RFREQ_TRAITVAL + XRCOV_TRAITVAL)/2)
of taxa with trait value
}
\description{
This internal function calculates metrics using traits having >2
  categories as values. Used for native status variables. Used in
  \code{calcNative()} function.
}
\section{Warning}{
 This function not intended for use on its own
}

\author{
Karen Blocksom
}
