% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VascPlantMetric_functions.r
\name{int.calcXBC}
\alias{int.calcXBC}
\title{Calculate mean Bray-Curtis distance among plots}
\usage{
int.calcXBC(x, sampID)
}
\arguments{
\item{x}{Data frame containing:

sampID: A character vector containing the name(s) of variable(s) necessary 
to identify unique samples

PLOT: Plot number of sample

SPECIES: Character code for taxon with no spaces

COVER: Percentage vegetative cover for a given taxon}

\item{sampID}{A character vector containing the name(s) of variable(s) 
necessary to identify unique samples}
}
\value{
Data frame consisting of \emph{sampID} variables and XCDIST_SPP, 
  defined as:
  
  Within AA dissimilarity based on species composition = Mean of between plot
  Bray-Curtis Distance (Dissimilarity)
}
\description{
This internal function calculates the mean Bray-Curtis distance 
  among plots based on species composition. Used in \code{calcBCmets()} 
  function.
}
\section{Warning}{
 This function not intended for use on its own
}

\author{
Karen Blocksom
}
