% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VascPlantMetric_functions.r
\name{prepareData}
\alias{prepareData}
\title{Prepare data for calculating metrics}
\usage{
prepareData(indf, inTaxa = taxaNWCA, inNatCC = ccNatNWCA, inWIS = wisNWCA)
}
\arguments{
\item{indf}{Data frame containing cover data summarized at
the UID, PLOT, and taxon value, with the following variable
names:
 \itemize{
\item UID: Unique sample identifier

\item PLOT: Plot number of data

\item USDA_NAME: Taxon name, based on USDA PLANTS names,
supplemented with NWCA names, if not in USDA PLANTS

\item COVER: Percent cover of taxon in plot, including zeros
for plots in which a taxon does not occur.
}}

\item{inTaxa}{Data frame with all taxa in indf, with variables:
\itemize{
\item USDA_NAME: Taxon name, based on USDA PLANTS names,
supplemented with NWCA names, if necessary.

\item FAMILY: Family name of taxon

\item GENUS: Genus name of taxon

\item CATEGORY (optional): USDA PLANTS category variable,
necessary to calculate category metrics.

\item DURATION (optional): USDA PLANTS duration variable,
necessary to calculate duration metrics.

\item GROWTH_HABIT (optional): USDA PLANTS growth habit
variable,necessary to calculate growth habit metrics

\item SPECIES_NAME_ID (optional): Taxonomic ID number,
which will be used in Bray-Curtis distance metrics
if available.

\item STATE: Postal abbreviation for state of sample

\item USAC_REGION: U.S. Army Corps of Engineers region
abbreviation for sample, to correspond to those
in inWIS data frame.
}}

\item{inNatCC}{Data frame with C-values and native status:

\itemize{
\item USDA_NAME: Taxon name

\item GEOG_ID: Postal abbreviation for STATE of taxon

\item NWCA_CC (optional): Coefficient of conservatism,
as used in NWCA, necessary to calculate CC metrics

\item NWCA_NATSTAT (optional): Native status variable,
as used in NWCA, necessary to calculate native status metrics.
}}

\item{inWIS}{Data frame with Wetland Indicator Status, from
U.S. Army Corps of Engineers (USAC):
\itemize{
\item USDA_NAME: Taxon name

\item GEOG_ID: USAC region, abbreviated to match those used in input
data frame

\item WIS: Wetland Indicator Status as provided by USAC
or added for NWCA
}}
}
\value{
A list containing six data frames:
\itemize{
 \item
   byUIDspp: Data frame with data summarized by UID and TAXON
   at the species level and contains:
   \itemize{
     \item UID: Unique sample identifier

     \item STATE: State abbreviation of site location, necessary for merging
with inNatCC

     \item USAC_REGION: U.S. Army Corps of Engineers region abbreviation
necessary for merging with inWIS.

     \item TAXON: Taxon name

     \item NUM: Number of occurrences of taxon across plots

     \item XABCOV: Mean percent absolute cover of taxon across plots

     \item DISTINCT: Distinctness value for each taxon, 1 if
the taxon has COVER>0 and 0 if not.

     \item NPLOTS: Number of plots in sample (1-5)

     \item TOTN: Total number of taxa in sample

     \item XTOTABCOV: Sum of XABCOV across all taxa in sample

     \item sXRCOV: taxon mean relative cover (XABCOV/XTOTABCOV)*100

     \item FREQ: Relative number of plots in which taxon occurs
(NUM/NPLOTS)*100

     \item TOTFREQ: Sum of FREQ across all taxa in sample

     \item SRFREQ: Relative frequency of taxon relative to total
frequency (FREQ/TOTFREQ)*100
   }
This data frame is also merged with the input taxa data
frames and contains, in addition, GENUS and FAMILY, but
also contains (depending on the input taxa lists): CATEGORY,
DURATION, GROWTH_HABIT, NWCA_CC, NWCA_NATSTAT, WIS.

\item byPlotspp: Data frame with data summarized by UID, PLOT, and
TAXON at the species level. Each data frame contains:
   \itemize{
     \item UID: Unique sample identifier

     \item PLOT: Plot number

     \item STATE: State of sample location

     \item USAC_REGION: USAC region code

     \item TAXON: Taxon name

     \item COVER: Sum of cover by TAXON within plot

     \item DISTINCT: Distinctness of taxon, value of 1 assigned to each row
}

\item byUIDgen: Data frame with data summarized by UID and TAXON
at the genus level and contains UID, STATE, USAC_REGION,
TAXON, NUM, XABCOV, and DISTINCT. NUM is the number of
plots in which taxon occurs, and XABCOV is the mean
absolute COVER across plots. DISTINCT is the value 1
assigned to each row.

\item byPlotgen: Data frame with data summarized by UID, PLOT, and
TAXON at the genus level. Each data frame contains UID,
PLOT, STATE, USAC_REGION, TAXON, COVER, and DISTINCT.
DISTINCT assigns the value for each taxon as 1 if the
taxon has COVER>0 or 0 if not. COVER is the sum of the
COVER variable.

\item byUIDfam: Data frame with data summarized by UID and TAXON
at the family level and contains UID, STATE, USAC_REGION,
TAXON, NUM, XABCOV, and DISTINCT. NUM is the number of plots
in which taxon occurs, and XABCOV is the mean absolute COVER
across plots. DISTINCT is the value 1 assigned to each row.

\item byPlotfam: Data frame with data summarized by UID, PLOT, and
TAXON at the family level. Each data frame contains UID,
PLOT, STATE, USAC_REGION, TAXON, COVER, and DISTINCT.
DISTINCT assigns the value for each taxon as 1 if the
taxon has COVER>0 or 0 if not. COVER is the sum of the
COVER variable.
}
}
\description{
Assemble datasets from data frame containing
vegetation cover by UID and PLOT, and taxa lists containing
variables used in metric calculation. Species-, genus-, and
family-level datasets are summarized by plot and UID. For the
species-level output data summarized by UID, various traits
are added to the output data set.
}
\details{
This function calls the createDFs() function, which
sums cover by UID, PLOT, TAXON, with sums > 100 truncated to
100 percent.
}
\examples{
head(VascPlantEx)
prepEx <- prepareData(VascPlantEx)

str(prepEx)
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}
\references{
US Environmental Protection Agency. 2016.
National Wetland Condition Assessment: 2011 Technical Report.
EPA-843-R-15-006. US Environmental Protection Agency,
Washington, DC.
}

