\name{Vtype_GrCovEx}
\alias{Vtype_GrCovEx}
\concept{Example Vegetation Type Data}
\concept{Example Ground Cover Data}
\docType{data}
\title{
Example Vegetation Type and Ground Surface Attributes data
}
\description{
A dataset containing vegetation type and ground surface attributes data for use in vtype_gcovMets() example.
}
\usage{data(Vtype_GrCovEx)}
\format{
  A data frame with 1775 observations on the following  variables.
  \describe{
    \item{UID}{A unique site visit ID}
    \item{PLOT}{Plot number of measurement}
    \item{PARAMETER}{Name of parameter measured}
    \item{RESULT}{Measured value of parameter}
  }
}
\note{
This dataset is a small subset of actual NWCA 2011 data, with UIDs recoded.
}
\examples{
data(Vtype_GrCovEx)
(Vtype_GrCovEx)
}
\keyword{datasets}
