% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wisNWCA_2016}
\alias{wisNWCA_2016}
\title{NWCA 2016 Wetland Indicator Status}
\format{
A data frame with 9584 observations on the following 5 variables.
\describe{
    \item{SPECIES_NAME_ID}{A numeric vector containing the NWCA taxonomic
    ID number}
    \item{NWCA_NAME}{Accepted USDA PLANTS name, or the official NWCA name
    where not available}
    \item{GEOG_ID}{Geographic unit to which value applies, which is US Army
    Corps region in this dataset}
    \item{WIS}{USAC-specific Wetland Indicator Status as used for NWCA 2016}
    \item{ECOIND1}{Wetland indicator status converted to numeric value, with
    lower numbers more characteristic of wetlands}
    \item{ECOIND2}{Wetland indicator status converted to numeric value, with
    higher numbers more characteristic of wetlands}
}
}
\usage{
wisNWCA_2016
}
\description{
A dataset containing NWCA 2016 USAC-specific Wetland Indicator Status (WIS)
values for vascular plant taxa
}
\note{
This dataset is the taxa list used for USAC-specific Wetland Indicator
Status values used in NWCA 2016.
}
\examples{
head(wisNWCA_2016)
str(wisNWCA_2016)

}
\keyword{datasets}
