% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegType_GrdCover_functions.r
\name{calcWcovMets}
\alias{calcWcovMets}
\title{Calculate water cover metrics}
\usage{
calcWcovMets(dataIn, nPlot)
}
\arguments{
\item{dataIn}{A data frame containing the following variables:
\itemize{
\item UID (unique sample identifier), PLOT (Sample plot from which data
were collected)

\item PARAMETER (specific measurement type)

\item RESULT (measured value)
}
The following parameters are used in
calculating tree metrics: 'TIME','MINIMUM_DEPTH',
'MAXIMUM_DEPTH','PREDOMINANT_DEPTH','TOTAL_WATER',
'WATER_NOVEG','WATER_AQVEG','WATER_EMERGVEG'.
Additional parameters or variables are ignored.}

\item{nPlot}{Number of plots sampled associated with each
UID with variables UID and NPLOTS}
}
\value{
Either a character string containing an error message
when metric calculation is not successful, or a data frame. Data
frame contains UID, PARAMETER, RESULT, where values of PARAMETER
consist of the metric name concatenated with trait value, with
valid values of: MIN_H2O_DEPTH, MAX_H2O_DEPTH,
XH2O_DEPTH_AA, MIN_COV_H2O, MAX_COV_H2O, FREQ_H2O,
FREQ_H2O_AQVEG, FREQ_H2O_EMERGVEG, FREQ_H2O_NOVEG, XCOV_H2O,
XCOV_H2O_AQVEG, XCOV_H2O_EMERGVEG, XCOV_H2O_NOVEG, IMP_H2O,
IMP_H2O_AQVEG, IMP_H2O_EMERGVEG, IMP_H2O_NOVEG, XH2O_DEPTH.
A list of metric descriptions is provided in the document
named VegTypes_GrdCover_Metric_Descriptions.pdf
included in the help directory for the package.
}
\description{
This internal function calculates metrics water
cover and depth based on ground cover data. This function is
called by calcVtype_GcovMets().
}
\details{
If any of the parameters are missing, they are assumed
to be zeros (if numeric), and metric values associated with any
metrics that cannot be calculated due to missing parameters are set
to a standardized value.
}
\examples{
head(Vtype_GrCovEx)
# Create data frame with number of plots sampled for each UID
nplots <- data.frame(UID=seq(1:10),NPLOTS=rep(5,10),stringsAsFactors=F)
# alternative approach to creating this data frame
nplots <- plyr::ddply(Vtype_GrCovEx,c('UID'),dplyr::summarise,
NPLOTS=length(unique(PLOT)))

wcovEx <- calcWcovMets(Vtype_GrCovEx,nplots)

head(wcovEx)
unique(wcovEx$METRIC)
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}

