% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWIS.r
\name{calcWIS}
\alias{calcWIS}
\title{Calculate Wetland Indicator Status metrics}
\usage{
calcWIS(vascIn, sampID = "UID")
}
\arguments{
\item{vascIn}{Data frame containing cover data summarized by
UID and TAXON, with the following fields:
\itemize{
    \item sampID: Variable(s) identified in \emph{sampID} argument

    \item TAXON: Taxon name

    \item WIS: Wetland Indicator Status, with possible values
    of FAC, FACU, FACW, OBL, UPL, NL, TBD, or missing.
    NL is recoded to UPL in function, TBD is recoded
    to missing in function.

    \item FREQ: Frequency of taxon among plots at site

    \item XABCOV: Mean percent cover of taxon across plots

    \item TOTN: Number of taxa in sample

    \item sXRCOV: proportion of summed cover across all taxa
    (XTOTABCOV) represented by taxon in sample

    \item NWCA_NATSTAT (optional): Native status variable with
      categories of 'NAT', 'ADV', 'CRYP', 'INTR', 'UND'.
      UND taxa are ignored.
   }}

\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples, 'UID' by default}
}
\value{
Data frame containing \emph{sampID} variables, PARAMETER, RESULT,
  where values of PARAMETER consist of the metric name concatenated with
  trait value (represented as TRAITNM below):
\itemize{
\item N_TRAITNM: Number of taxa with trait

\item PCTN_TRAITNM: Number of taxa with trait as percentage of \emph{TOTN}

\item XABCOV_TRAITNM: Sum of \emph{XABCOV} values across taxa with trait

\item XRCOV_TRAITNM: Sum of \emph{sXRCOV} values across taxa with trait
}
In addition WIS indices based on all and native species only (if
NWCA_NATSTAT is provided in the input data frame), with the
suffixes ALL and NAT, respectively. WIS values recoded as numeric
with OBL=1, FACW=2, FAC=3, FACU=4, UPL=5 for WETIND metrics and as
OBL=5, FACW=4, FAC=3 FACU=2, and UPD=1 for WETIND2 metrics:
\itemize{ \item WETIND_COV_ALL, WETIND_COV_NAT: Wetland Index, weighted by
taxon cover, based on numeric conversion of WIS, with lower numbers 
indicating wetter conditions

\item WETIND_FREQ_ALL, WETIND_FREQ_NAT: Wetland Index, weight by frequency, 
based on numeric conversion of WIS, with lower numbers 
indicating wetter conditions } 

\itemize{ \item WETIND2_COV_ALL, WETIND2_COV_NAT: Wetland Index, weighted by
taxon cover, based on numeric conversion of WIS, with higher numbers 
indicating wetter conditions

\item WETIND2_FREQ_ALL, WETIND2_FREQ_NAT: Wetland Index, weight by frequency, 
based on numeric conversion of WIS, with higher numbers 
indicating wetter conditions } 

If NWCA_NATSTAT is provided in the input data frame, the following metrics
are calculated based on Alien + Cryptogenic species: 
\itemize{ \item
N_OBLFACW_AC: Number of alien and cryptogenic taxa with WIS of either OBL or
FACW.

\item XABCOV_OBLFACW_AC: Mean absolute cover of alien and cryptogenic taxa 
with WIS of either OBL or FACW.

\item XRCOV_OBLFACW_AC: Mean relative cover of alien and cryptogenic taxa 
with WIS of either OBL or FACW. }
}
\description{
This function calculates Wetland Indicator Status (WIS)
metrics, including variations based on native status,
if NWCA_NATSTAT is present in the input data frame.
}
\examples{
head(VascPlantEx)
 exPlant <- prepareData(VascPlantEx, taxon_name = 'USDA_NAME', 
 inTaxa = taxaNWCA, inNat = ccNatNWCA, inCVal = ccNatNWCA, 
 inWIS = wisNWCA, cValReg='STATE')

wisEx <- calcWIS(exPlant$byUIDspp)

head(wisEx)
unique(wisEx$PARAMETER)
}
\references{
US Environmental Protection Agency. 2016. National Wetland
Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US
Environmental Protection Agency, Washington, DC.
}
\author{
Karen Blocksom \email{Blocksom.karen@epa.gov}
}
