% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VascPlantMetric_functions.r
\name{createDFs}
\alias{createDFs}
\title{Create data frames for function input}
\usage{
createDFs(
  sampID = "UID",
  tvar,
  vascIn,
  taxon_name,
  taxa,
  state = "STATE",
  coeReg = "USAC_REGION",
  cValReg = "NWC_CREG"
)
}
\arguments{
\item{sampID}{A character vector containing the name(s) of
variable(s) necessary to identify unique samples, 'UID' by default}

\item{tvar}{String with the level of taxonomy
(taxon_name,'GENUS','FAMILY')}

\item{vascIn}{Data frame with vegetation cover data, having
the following variables:
\itemize{

\item sampID: Variable(s) found in the argument \emph{sampID}

\item PLOT:  Plot number of data (1 to 5 possible)

\item Variable named in \emph{state}: Two-letter state postal code of site, used to link
native status to taxa in native status taxalist (inNat)

\item Variable named in \emph{coeReg}: U.S. Army Corps of Engineers region abbreviation for
  sample, to correspond to GEOG_ID in Wetland Indicator Status taxalist (inWIS)

\item Variable named in \emph{cValReg}: NWCA C-value region abbreviation for sample, to correspond
those in C-value taxalist

\item Variable named in \emph{taxon_name}: Taxon name, must match with taxa data frame

\item COVER: Percentage estimated vegetation cover by TAXON in PLOT
}}

\item{taxon_name}{String containing the name of variable for taxon name in 
\emph{vascIn} and in taxalists.}

\item{taxa}{Data frame containing USDA_NAME, GENUS, CATEGORY,
GROWTH_HABIT, and DURATION variables. Dataset taxaNWCA is the
default if not specified. These variables are assumed to be
populated with abbreviations as found in USDA PLANTS database.}

\item{state}{String containing the name of the state in \emph{vascIn}, 
with default value of 'STATE'}

\item{coeReg}{String containing the name of the U.S. Army Corps of Engineers
region in \emph{vascIn} associated with Wetland Indicator Status, 
with default value of 'USAC_REGION'}

\item{cValReg}{String containing the name of the variable in \emph{vascIn}
which specifies the C-value region.}
}
\value{
A list containing two data frames. The first data frame summarizes
  data by \emph{sampID} variables and TAXON and contains sampID variables,
  STATE, USAC_REGION, TAXON, NUM, XABCOV, DISTINCT, and NPLOTS. NUM is the
  number of plots in which taxon occurs, and XABCOV is the mean absolute
  COVER across plots. DISTINCT is the value 1 assigned to each row. NPLOTS is
  the number of plots sampled.
  
  The second summarizes by \emph{sampID} variables, PLOT, and TAXON. Each
  data frame contains sampID variables, PLOT, STATE, USAC_REGION, TAXON,
  COVER, and DISTINCT. DISTINCT assigns the value for each taxon as 1 if the 
  taxon has COVER>0 or 0 if not. COVER is the sum of the COVER variable by
  sampID variables, PLOT, and TAXON.
}
\description{
This internal function merges the inputs of
plant cover data and taxalist and summarizes by the taxonomic
level specified. Not intended for use on its own.
}
\examples{
head(VascPlantEx)
data(taxaNWCA)

outEx <- createDFs(sampID='UID','GENUS',VascPlantEx,taxon_name = 'USDA_NAME',
taxaNWCA,cValReg='STATE')
head(outEx$byUID)
head(outEx$byPlot)
}
\author{
Karen Blocksom \email{blocksom.karen@epa.gov}
}
