% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{VascPlantEx}
\alias{VascPlantEx}
\title{Example Vascular Plant Cover data}
\format{A data frame with 1200 observations on the following 6 variables.
  \describe{
  \item{UID}{A unique site visit ID}
  \item{STATE}{Two letter state code of sample site}
  \item{USAC_REGION}{US Army Corps region of sample site}
  \item{PLOT}{Plot number of measurement}
  \item{USDA_NAME}{Taxon name corresponding to input taxa tables}
  \item{COVER}{A numeric vector representing the percent cover of a 
  taxon within a plot}
  }}
\usage{
VascPlantEx
}
\description{
A dataset containing vascular plant data for use in vascPlantMets() example.
}
\note{
This dataset is a small subset of actual NWCA 2011 data, with UIDs 
recoded.
}
\examples{
head(VascPlantEx)
str(VascPlantEx) 
}
\keyword{datasets}
