% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Vtype_GrCovEx}
\alias{Vtype_GrCovEx}
\title{Example Vegetation Type and Ground Surface Attributes data}
\format{
A data frame with 1775 observations on the following  variables.
 \describe{
 \item{UID}{A unique site visit ID}
 \item{PLOT}{Plot number of measurement}
 \item{PARAMETER}{Name of parameter measured}
 \item{RESULT}{Measured value of parameter}
 }
}
\usage{
Vtype_GrCovEx
}
\description{
A dataset containing vegetation type and ground surface attributes data
for use in vtype_gcovMets() example.
}
\note{
This dataset is a small subset of actual NWCA 2011 data, with UIDs
 recoded.
}
\examples{
head(Vtype_GrCovEx)
str(Vtype_GrCovEx)
}
\keyword{datasets}
