% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wisNWCA}
\alias{wisNWCA}
\title{NWCA 2011 Wetland Indicator Status}
\format{
A data frame with 19818 observations on the following 5 variables.
\describe{
    \item{SPECIES_NAME_ID}{A numeric vector containing the NWCA taxonomic
    ID number}
    \item{USDA_NAME}{Accepted USDA PLANTS name, or the official NWCA name
    where not available}
    \item{GEOG_ID}{Geographic unit to which value applies, which is US Army
    Corps region in this dataset}
    \item{WIS}{USAC-specific Wetland Indicator Status as used for NWCA 2011}
}
}
\usage{
wisNWCA
}
\description{
A dataset containing NWCA 2011 USAC-specific Wetland Indicator Status (WIS)
values for vascular plant taxa
}
\note{
This dataset is the taxa list used for USAC-specific Wetland Indicator
Status values used in NWCA 2011.
}
\examples{
head(wisNWCA)
str(wisNWCA)

}
\keyword{datasets}
